run to launch in legacy mode
