+++ python/paddle/distributed/launch/context/device.py
