+++ python/paddle/distributed/launch/context/resource.py
